Initial Implementation of component system
Implement the component system, consisting of a component registry, a dependency
injection library and a mechanism for scoping values. This enables decoupled and
declarative programming styles in application writing.

This is designed to operate in the context of frameworks for building e.g. web
applications and is fed by a build tool, both of which are under development. It
is being committed in a threadbare state (not much docs or tests) to accelerate
those projects.